chore: relocate commands/utils
