+++ packages/esw/src/utils/data-structure.ts
@@ -1,0 +1,10 @@
+// https://github.com/microsoft/TypeScript/pull/39094#Spreads_in_array_literals
+export function tuple<T extends unknown[]>(args: [...T]) {
+  return args
+}
+
+export function isFulfillResult<Data>(
+  result: PromiseSettledResult<Data>
+): result is PromiseFulfilledResult<Data> {
+  return result.status === 'fulfilled'
+}
