+++ packages/esw/src/utils/env.ts
@@ -1,0 +1,3 @@
+export function isProduction(process: NodeJS.Process) {
+  return process.env['NODE_ENV'] === 'production'
+}
